pyproject.toml
[project]
name = "periodic-scraper"
version = "0.1.0"
description = "webscraper for news and articles"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "huggingface>=0.0.1",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.9.7",
    "lxml>=4.9.2"
]
